Ecg Recordings Updated and Minor Model Changes

+++ src/main/kotlin/com/ecgapp/ecgapp/dto/medicalCond.kt
@@ -3,7 +3,7 @@ package com.ecgapp.ecgapp.dto
 import java.time.LocalDate
 
 data class MedicalConditionInput(
-    val medicalInfoId: Int,
+    val medicalInfoId: Long,
     val conditionName: String,
     val severity: String,
     val diagnosedDate: String,
