+++ src/main/kotlin/com/ecgapp/ecgapp/repository/MedicalInfoRepository.kt
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface MedicalInfoRepository : JpaRepository<MedicalInfo, Int> {
+interface MedicalInfoRepository : JpaRepository<MedicalInfo, Long> {
     fun findByUserId(userId: Long): MedicalInfo?
 }
